requirements.txt
streamlit
sentence-transformers
pdfplumber
python-docx
google-generativeai
scikit-learn
python-dotenv
